str to int

+++ bot/bot.py
@@ -7,8 +7,8 @@ from urllib.request import urlopen
 import discord
 
 DISCORD_TOKEN = getenv("DISCORD_TOKEN")
-MAX_IMAGE_SIZE_MB = getenv("MAX_IMAGE_SIZE_MB")
-MAX_IMAGE_SIZE_BYTES = MAX_IMAGE_SIZE_MB * 1000 * 1000
+MAX_IMAGE_SIZE_MB = int(getenv("MAX_IMAGE_SIZE_MB"))
+MAX_IMAGE_SIZE_BYTES = int(MAX_IMAGE_SIZE_MB * 1000 * 1000)
 client = discord.Client()
